+++ training_log_app/src/repositories/theme_repository.py
@@ -1,39 +1,36 @@
 from src.database_connection import get_database_connection
 
+
 class ThemeRepository:
     '''Class to handle saving
     and retrieving themes from database.
+    Themes style the user interface.
     '''
 
     def __init__(self, db_connection):
         self._db_connection = db_connection
 
-
-    
     def load_application_theme(self):
         '''Gets apps theme from database.
-        Has nothing to do with user but this is
-        simpler than creating new repository.
-        
+
         Returns:
             string: theme
         '''
 
         cursor = self._db_connection.cursor()
-        
+
         theme = cursor.execute('''select theme from Themes
                                 where id=1''').fetchone()
-        
+
         self._db_connection.commit()
 
         return theme[0]
-    
 
     def save_application_theme(self, theme: str):
         '''Save current theme in the database to be
         loaded the next time. First deletes previously
         saved theme.
-        
+
         Args:
             theme: theme to be saved
         '''
@@ -44,7 +41,7 @@ class ThemeRepository:
 
         cursor.execute('''insert into Themes (theme)
                         values (?)''', [theme])
-        
+
         self._db_connection.commit()