+++ training_log_app/src/ui/ui.py
@@ -7,6 +7,7 @@ from ui.statistics_view import StatisticsPageView
 from ui.browse_log_entries_view import BrowseLogEntriesView
 from services.log_entry_service import log_entry_service
 
+
 class UI:
     '''Class responsible for user interface'''
 
@@ -18,34 +19,39 @@ class UI:
 
         self._root = root
         self._current_view = None
-        
+
         self._theme = log_entry_service.load_theme()
 
         self.configure_window()
-        
+
     def configure_window(self):
         self._root.resizable(width=False, height=False)
-        
+
     def _hide_current_view(self):
         if self._current_view:
             self._current_view.destroy()
-        
+
     def _change_theme(self):
         '''Changes applications theme.
         '''
-        if self._theme == 'day':
-            self._theme = 'night'
+        if self._theme == 'DayMode':
+            self._theme = 'NightMode'
         else:
-            self._theme = 'day'
+            self._theme = 'DayMode'
 
         log_entry_service.save_theme(current_theme=self._theme)
         self._show_main_user_page()
 
     def start(self):
-        '''Starts ui'''
+        '''Starts ui.
+        '''
+
         self._show_login_view()
 
     def _show_new_log_entry_view(self):
+        '''Displays new log entry creation page.
+        '''
+
         self._hide_current_view()
 
         self._current_view = NewLogEntryView(self._root,
@@ -55,6 +61,9 @@ class UI:
         self._current_view.pack()
 
     def _show_login_view(self):
+        '''Displays login page.
+        '''
+
         self._hide_current_view()
 
         self._current_view = LoginView(self._root,
@@ -65,6 +74,9 @@ class UI:
         self._current_view.pack()
 
     def _show_main_user_page(self):
+        '''Displays main page.
+        '''
+
         self._hide_current_view()
         self._current_view = MainUserPageView(self._root,
                                               self._show_login_view,
@@ -76,6 +88,9 @@ class UI:
         self._current_view.pack()
 
     def _show_create_new_user_view(self):
+        '''Displays new user creation page.
+        '''
+        
         self._hide_current_view()
 
         self._current_view = CreateNewUserView(self._root,
@@ -85,6 +100,9 @@ class UI:
         self._current_view.pack()
 
     def _show_statistics_page_view(self):
+        '''Displays statistics page.
+        '''
+
         self._hide_current_view()
 
         self._current_view = StatisticsPageView(self._root,
@@ -94,6 +112,9 @@ class UI:
         self._current_view.pack()
 
     def _show_browse_log_entries_view(self):
+        '''Displays browsing log entries page.
+        '''
+        
         self._hide_current_view()
 
         self._current_view = BrowseLogEntriesView(self._root,