+++ training_log_app/tasks.py
@@ -1,5 +1,7 @@
 from invoke import task
 
+'''Tasks to be called by poetry invoke.
+'''
 
 @task
 def start(ctx):